refactoring story docs
