+++ packages/rhdf-demo/.storybook/config.js
@@ -1,35 +1,42 @@
 import { configure } from "@storybook/react";
-import { addDecorator } from "@storybook/react";
-import { withInfo } from "@storybook/addon-info";
 
-// Show story info (source code ...)
-addDecorator(
-  withInfo({
-    styles: {
-      header: {
-        h1: {
-          marginRight: "20px",
-          fontSize: "25px",
-          display: "inline",
-        },
-        body: {
-          paddingTop: 0,
-          paddingBottom: 0,
-        },
-        h2: {
-          display: "inline",
-          color: "#999",
-        },
-      },
-      infoBody: {
-        padding: 0,
-        lineHeight: "2",
-      },
-    },
-    inline: true,
-    source: true,
-  })
-);
+/**
+ * Addon-info is being replaced by addon-docs docsPage
+ * https://github.com/storybookjs/storybook/blob/master/addons/docs/docs/docspage.md
+ */
+// import { addDecorator } from "@storybook/react";
+// import { withInfo } from "@storybook/addon-info";
+
+// Add default info style, hide all modules and show if needed in stories
+// addDecorator(
+//   withInfo({
+//     styles: {
+//       header: {
+//         h1: {
+//           marginRight: "20px",
+//           fontSize: "25px",
+//           display: "inline",
+//         },
+//         body: {
+//           paddingTop: 0,
+//           paddingBottom: 0,
+//         },
+//         h2: {
+//           display: "inline",
+//           color: "#999",
+//         },
+//       },
+//       infoBody: {
+//         padding: 0,
+//         lineHeight: "2",
+//       },
+//     },
+//     inline: true,
+//     header: false,
+//     source: false,
+//     propTables: false,
+//   })
+// );
 
 // automatically import all files ending in mdx tsx
 configure(require.context("../src", true, /\.stories\.(mdx|[tj]sx?)$/), module);
