+++ packages/rhdf-demo/package.json
@@ -41,7 +41,6 @@
     "@babel/core": "^7.7.5",
     "@storybook/addon-actions": "^5.2.8",
     "@storybook/addon-docs": "^5.2.8",
-    "@storybook/addon-info": "^5.2.8",
     "@storybook/addon-links": "^5.2.8",
     "@storybook/addons": "^5.2.8",
     "@storybook/preset-create-react-app": "^1.3.2",
